Problem 4 finished
This is my working solution to problem 4, though not the most elegant.
The only help I received was to get the actual answer when I finally
got the findLargestPal function to work properly, but wasn't getting
the right answer. I sort of explain how I solved that problem on line
69. I did not look at any outside code.